Added a shotgun, and a pause menu
Press 1 to switch weapon and press escape to go to pause menu. Use up
down arrow to switch buttons and enter to pick your button. Got pissed
with the moving thing so i added a circle on the camera and made gun
shoot a raycast from middle of camera instead of where your mouse is.
You can change that if you want.